version 2.4.0
* support for the ARRL June VHF contest
* support for reading ADIF qso information via UDP from WSJT-X
* allow per-band setting of IF offset and inversion (same as swap IQ) in so2sdr-
bandmap
* To support higher frequency bands, the  internal format for frequency stored i
n log was changed.

+++ so2sdr-bandmap/bandmapentry.h
@@ -30,7 +30,7 @@ public:
     BandmapEntry();
 
     bool       dupe;
-    int        f;
+    double     f;
     qint64     createdTime;
     QByteArray call;
 };
